ci: pin the linter, clear the backlog, and stop lint from hiding the tests
The Tests workflow has been failing on lint for a long time, and because
Lint and Tests were steps of the SAME job, the failure meant `pytest`
never ran at all. The badge said "failing" and nobody could tell whether
that meant a style nit or a broken subnet.

Root cause: the dev extra asked for `ruff>=0.1`, so CI resolved whatever
had shipped that week. Its rule set is not stable across releases, so the
check drifted away from anything a contributor could reproduce — the same
tree reported 42 findings on one version and 804 on another.

- ruff is pinned in the dev extra, with a note on why bumping it is a
  deliberate act.
- The backlog is cleared: 804 -> 0. Most were mechanical (modern optional
  syntax, import order, unused imports, explicit subprocess check
  arguments, context managers for file reads, executable bits on files
  that carry a shebang).
- Four rules are switched off in pyproject with the reason written down
  next to each: broad excepts at the chain and container boundaries, the
  two skip-on-error patterns that go with them, and naive UTC datetimes,
  which are a data-model decision rather than a lint fix. B008 is
  configured rather than ignored — every instance was either FastAPI's
  dependency idiom or a frozen dataclass, so the shared-mutable-default
  hazard it warns about cannot occur.
- Lint and Tests are now separate jobs, and the test job does not depend
  on the lint job, so neither can hide the other again. The matrix no
  longer cancels sibling interpreters on the first failure.

The suite now passes: 1,577 tests, 0 failures.

The six long-standing failures were not six problems. They were one: when
the flat-week fallback shipped, only one of the two copies of the tiered
allocator's tests was updated. The tests that assert the participation
GATE now disable the top-up explicitly, which is what they were always
trying to express; the small-pool test was rewritten to the banded
behaviour that replaced the single-pool collapse; and the stale duplicate
of tests/unit/validator/test_tiered_weights.py was deleted, its maintained
twin covering the same ground and more.

+++ hope/commitment/prediction_payload.py
@@ -50,8 +50,8 @@ import os
 from dataclasses import dataclass
 from typing import Any
 
-from cryptography.hazmat.primitives.ciphers.aead import AESGCM
 from cryptography.hazmat.primitives.asymmetric.ed25519 import Ed25519PrivateKey
+from cryptography.hazmat.primitives.ciphers.aead import AESGCM
 
 from hope.commitment.canonical import (
     aes_gcm_aad,
@@ -59,7 +59,6 @@ from hope.commitment.canonical import (
     canonical_cbor_loads,
 )
 from hope.commitment.inner_sig import add_inner_sig, verify_inner_sig
-
 
 # Protocol version embedded in every payload.
 PREDICTION_PAYLOAD_VERSION = 1
@@ -98,7 +97,7 @@ def _scale_quantile(value: float) -> int:
 
     -3.7% → -37; +12.5% → 125. Rounded half-to-even (Python default).
     """
-    return int(round(value * QUANTILE_SCALE))
+    return round(value * QUANTILE_SCALE)
 
 
 def _scale_probability(value: float) -> int:
@@ -108,7 +107,7 @@ def _scale_probability(value: float) -> int:
     """
     if not (0.0 <= value <= 1.0):
         raise ValueError(f"probability out of range [0, 1]: {value}")
-    return int(round(value * PROBABILITY_SCALE))
+    return round(value * PROBABILITY_SCALE)
 
 
 def _validate_quantile_triple(q: tuple[float, float, float], name: str) -> list[int]:
